Move availability handler into UWS library
Currently, the /availability route only uses UWS information, so it
can be moved entirely into the UWS library. If this needs to be more
complicated in the future, we can revisit it.

Use vo-cutouts for the Availability model and XML rendering instead
of templating.
